--- src/app/tap/components/RhythmBar/style.module.scss
@@ -1,28 +1,0 @@
-.container {
-	border-radius: 4px;
-	overflow: hidden;
-}
-  
-.progressContainer {
-	width: 100%;
-	background-color: #eee;
-	height: 20px;
-	display: flex;
-	position: relative;
-}
-  
-.progressBar {
-	height: 100%;
-	transition: width 0.1s ease-in-out;
-	position: absolute; // 進捗バーとハイライト区間の重ね表示のため
-	top: 0;
-	left: 0;
-}
-  
-.highlightedSection {
-	height: 100%;
-	position: absolute;
-	z-index: 1;
-}
-  
-
