+++ src/app/tap/features/RecordRhythm/components/RhythmBar/index.tsx
@@ -3,6 +3,7 @@
 import { useState, useEffect } from "react";
 import styles from "./style.module.scss";
 import colors from "@/styles/colors.module.scss";
+import { useRhythmBar } from "../../hooks/useRhythmBar";
 
 type ProgressBarProps = {
   duration: number; // バーが満タンになるまでの時間（秒）
@@ -10,43 +11,8 @@ type ProgressBarProps = {
   barWidth:number;
 };
 
-export default function ProgressBar({ duration, highlight = false ,barWidth}: ProgressBarProps) {
-  const [progress, setProgress] = useState(0);
-  const [highlightedSections, setHighlightedSections] = useState<{ start: number, end: number }[]>([]);
-  const [currentHighlightStart, setCurrentHighlightStart] = useState<number | null>(null);
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setProgress((prev) => {
-        const increment = 100 / (duration * 100);
-        const newProgress = parseFloat((prev + increment).toFixed(2));
-        if (newProgress >= 100) {
-          clearInterval(interval);
-          return 100;
-        }
-        return newProgress;
-      });
-    }, 50);
-
-    return () => clearInterval(interval);
-  }, [duration]);
-
-  useEffect(() => {
-    if (highlight) {
-      if (currentHighlightStart === null && progress < 100) {
-        setCurrentHighlightStart(progress);
-      }
-    } else {
-      if (currentHighlightStart !== null) {
-        setHighlightedSections((prev) => [
-          ...prev,
-          { start: currentHighlightStart, end: progress },
-        ]);
-        setCurrentHighlightStart(null);
-      }
-    }
-  }, [progress, highlight]);
-
+export default function ProgressBar({ duration,barWidth}: ProgressBarProps) {
+  const { pushSpaceKey,progress,highlightedSections,currentHighlightStart} = useRhythmBar(duration)
   // ハイライト区間の表示
   const renderHighlightedSections = () => {
     return highlightedSections.map((section, index) => (
@@ -64,10 +30,10 @@ export default function ProgressBar({ duration, highlight = false ,barWidth}: Pr
 
   return (
     <div 
-	  className={styles.container}
-	  style={{
+    className={styles.container}
+    style={{
 		width: `${barWidth}%`
-	  }}
+    }}
 	>
       <div className={styles.progressContainer}>
         {/* 進捗バー */}
@@ -92,7 +58,7 @@ export default function ProgressBar({ duration, highlight = false ,barWidth}: Pr
           />
 
           {/* ハイライト区間の表示 */}
-          {highlight && currentHighlightStart !== null && (
+          {pushSpaceKey && currentHighlightStart !== null && (
             <div
               className={styles.highlightedSection}
               style={{
